requirements.txt
matplotlib==3.6.1
numpy==1.23.4
Pillow==9.3.0
torch==1.10.0+cu102
torchsummary==1.5.1
tqdm==4.64.1
